+++ cmd/http-check/main_test.go
@@ -43,7 +43,7 @@ func TestExecuteCheck(t *testing.T) {
 		plugin.SearchString = tc.search
 		status, err := executeCheck(event)
 		assert.NoError(err)
-		assert.Equal(status, tc.status)
+		assert.Equal(tc.status, status)
 	}
 
 	testCasesStatus := []struct {
@@ -79,6 +79,22 @@ func TestExecuteCheck(t *testing.T) {
 		plugin.RedirectOK = tc.allowRedirect
 		status, err := executeCheck(event)
 		assert.NoError(err)
-		assert.Equal(status, tc.returnStatus)
+		assert.Equal(tc.returnStatus, status)
 	}
+
+	event := corev2.FixtureEvent("entity1", "check")
+	assert := assert.New(t)
+
+	var test = httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		assert.Equal("Test Header 1 Value", r.Header.Get("Test-Header-1"))
+		assert.Equal("Test Header 2 Value", r.Header.Get("Test-Header-2"))
+	}))
+	_, err := url.ParseRequestURI(test.URL)
+	require.NoError(t, err)
+	plugin.URL = test.URL
+	plugin.SearchString = ""
+	plugin.Headers = []string{"Test-Header-1: Test Header 1 Value", "Test-Header-2: Test Header 2 Value"}
+	status, err := executeCheck(event)
+	assert.NoError(err)
+	assert.Equal(sensu.CheckStateOK, status)
 }