+++ cmd/http-json/main_test.go
@@ -70,4 +70,22 @@ func TestExecuteCheck(t *testing.T) {
 		assert.NoError(err)
 		assert.Equal(tc.status, status)
 	}
+
+	assert := assert.New(t)
+
+	var test = httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		assert.Equal("Test Header 1 Value", r.Header.Get("Test-Header-1"))
+		assert.Equal("Test Header 2 Value", r.Header.Get("Test-Header-2"))
+		w.WriteHeader(http.StatusOK)
+		_, _ = w.Write(testJSON)
+	}))
+	_, err := url.ParseRequestURI(test.URL)
+	require.NoError(t, err)
+	plugin.URL = test.URL
+	plugin.Path = "number"
+	plugin.Expression = "== 10"
+	plugin.Headers = []string{"Test-Header-1: Test Header 1 Value", "Test-Header-2: Test Header 2 Value"}
+	status, err := executeCheck(event)
+	assert.NoError(err)
+	assert.Equal(sensu.CheckStateOK, status)
 }